+++ packages/apidom-reference/src/dereference/strategies/openapi-2/visitor.ts
@@ -12,6 +12,7 @@ import {
   cloneDeep,
   toValue,
   Namespace,
+  ObjectElement,
 } from '@swagger-api/apidom-core';
 import { ApiDOMError } from '@swagger-api/apidom-error';
 import { evaluate, URIFragmentIdentifier } from '@swagger-api/apidom-json-pointer/modern';
@@ -105,6 +106,35 @@ class OpenAPI2DereferenceVisitor {
     this.refractCache = refractCache;
   }
 
+  protected handleDereferenceError(error: unknown, refEl: Element) {
+    if (this.options.dereference.dereferenceOpts?.continueOnError) {
+      this.options.dereference.dereferenceOpts?.errors.push({ error, refEl });
+      return undefined;
+    }
+    throw error;
+  }
+
+  protected getNestedVisitorOptions(referencingElement: ObjectElement): ReferenceOptions {
+    return {
+      ...this.options,
+      resolve: {
+        ...this.options.resolve,
+        external:
+          this.options.dereference?.dereferenceOpts?.skipNestedExternal &&
+          toValue(referencingElement.get('$ref')).startsWith('#')
+            ? false
+            : this.options.resolve.external,
+      },
+      dereference: {
+        ...this.options.dereference,
+        dereferenceOpts: {
+          ...this.options.dereference?.dereferenceOpts,
+          continueOnError: false,
+        },
+      },
+    };
+  }
+
   protected toBaseURI(uri: string): string {
     return url.resolve(this.reference.uri, url.sanitize(url.stripHash(uri)));
   }
@@ -194,15 +224,30 @@ class OpenAPI2DereferenceVisitor {
       return false;
     }
 
-    const reference = await this.toReference(toValue(referencingElement.$ref));
+    let reference: Reference;
+
+    try {
+      reference = await this.toReference(toValue(referencingElement.$ref));
+    } catch (error) {
+      return this.handleDereferenceError(error, referencingElement);
+    }
+
     const $refBaseURI = url.resolve(retrievalURI, toValue(referencingElement.$ref));
 
     this.indirections.push(referencingElement);
 
     const jsonPointer = URIFragmentIdentifier.fromURIReference($refBaseURI);
 
-    // possibly non-semantic fragment
-    let referencedElement = evaluate<Element>(reference.value.result, jsonPointer);
+    let referencedElement: Element;
+
+    try {
+      // possibly non-semantic fragment
+      referencedElement = evaluate<Element>(reference.value.result, jsonPointer);
+    } catch (error) {
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
+    }
+
     referencedElement.id = identityManager.identify(referencedElement);
 
     /**
@@ -229,14 +274,18 @@ class OpenAPI2DereferenceVisitor {
 
     // detect direct or circular reference
     if (referencingElement === referencedElement) {
-      throw new ApiDOMError('Recursive Reference Object detected');
+      const error = new ApiDOMError('Recursive Reference Object detected');
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
     }
 
     // detect maximum depth of dereferencing
     if (this.indirections.length > this.options.dereference.maxDepth) {
-      throw new MaximumDereferenceDepthError(
+      const error = new MaximumDereferenceDepthError(
         `Maximum dereference depth of "${this.options.dereference.maxDepth}" has been exceeded in file "${this.reference.uri}"`,
       );
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
     }
 
     // detect second deep dive into the same fragment and avoid it
@@ -244,8 +293,12 @@ class OpenAPI2DereferenceVisitor {
       reference.refSet!.circular = true;
 
       if (this.options.dereference.circular === 'error') {
-        throw new ApiDOMError('Circular reference detected');
-      } else if (this.options.dereference.circular === 'replace') {
+        const error = new ApiDOMError('Circular reference detected');
+        this.indirections.pop();
+        return this.handleDereferenceError(error, referencingElement);
+      }
+
+      if (this.options.dereference.circular === 'replace') {
         const refElement = new RefElement(referencedElement.id, {
           type: 'reference',
           uri: reference.uri,
@@ -277,7 +330,8 @@ class OpenAPI2DereferenceVisitor {
       (isExternalReference ||
         isNonEntryDocument ||
         isReferenceElement(referencedElement) ||
-        shouldDetectCircular) &&
+        shouldDetectCircular ||
+        this.options.dereference.dereferenceOpts?.continueOnError) &&
       !ancestorsLineage.includesCycle(referencedElement)
     ) {
       // append referencing reference to ancestors lineage
@@ -287,14 +341,19 @@ class OpenAPI2DereferenceVisitor {
         reference,
         namespace: this.namespace,
         indirections: [...this.indirections],
-        options: this.options,
+        options: this.getNestedVisitorOptions(referencingElement),
         refractCache: this.refractCache,
         ancestors: ancestorsLineage,
       });
-      referencedElement = await visitAsync(referencedElement, visitor, {
-        keyMap,
-        nodeTypeGetter: getNodeType,
-      });
+      try {
+        referencedElement = await visitAsync(referencedElement, visitor, {
+          keyMap,
+          nodeTypeGetter: getNodeType,
+        });
+      } catch (error) {
+        this.indirections.pop();
+        return this.handleDereferenceError(error, referencingElement);
+      }
 
       // remove referencing reference from ancestors lineage
       directAncestors.delete(referencingElement);
@@ -367,15 +426,29 @@ class OpenAPI2DereferenceVisitor {
       return undefined;
     }
 
-    const reference = await this.toReference(toValue(referencingElement.$ref));
+    let reference: Reference;
+
+    try {
+      reference = await this.toReference(toValue(referencingElement.$ref));
+    } catch (error) {
+      return this.handleDereferenceError(error, referencingElement);
+    }
+
     const $refBaseURI = url.resolve(retrievalURI, toValue(referencingElement.$ref));
 
     this.indirections.push(referencingElement);
 
     const jsonPointer = URIFragmentIdentifier.fromURIReference($refBaseURI);
 
-    // possibly non-semantic referenced element
-    let referencedElement = evaluate<Element>(reference.value.result, jsonPointer);
+    let referencedElement: Element;
+
+    try {
+      // possibly non-semantic referenced element
+      referencedElement = evaluate<Element>(reference.value.result, jsonPointer);
+    } catch (error) {
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
+    }
     referencedElement.id = identityManager.identify(referencedElement);
 
     /**
@@ -394,14 +467,18 @@ class OpenAPI2DereferenceVisitor {
 
     // detect direct or indirect reference
     if (referencingElement === referencedElement) {
-      throw new ApiDOMError('Recursive Path Item Object reference detected');
+      const error = new ApiDOMError('Recursive Path Item Object reference detected');
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
     }
 
     // detect maximum depth of dereferencing
     if (this.indirections.length > this.options.dereference.maxDepth) {
-      throw new MaximumDereferenceDepthError(
+      const error = new MaximumDereferenceDepthError(
         `Maximum dereference depth of "${this.options.dereference.maxDepth}" has been exceeded in file "${this.reference.uri}"`,
       );
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
     }
 
     // detect second deep dive into the same fragment and avoid it
@@ -409,8 +486,12 @@ class OpenAPI2DereferenceVisitor {
       reference.refSet!.circular = true;
 
       if (this.options.dereference.circular === 'error') {
-        throw new ApiDOMError('Circular reference detected');
-      } else if (this.options.dereference.circular === 'replace') {
+        const error = new ApiDOMError('Circular reference detected');
+        this.indirections.pop();
+        return this.handleDereferenceError(error, referencingElement);
+      }
+
+      if (this.options.dereference.circular === 'replace') {
         const refElement = new RefElement(referencedElement.id, {
           type: 'path-item',
           uri: reference.uri,
@@ -442,7 +523,8 @@ class OpenAPI2DereferenceVisitor {
       (isExternalReference ||
         isNonEntryDocument ||
         (isPathItemElement(referencedElement) && isStringElement(referencedElement.$ref)) ||
-        shouldDetectCircular) &&
+        shouldDetectCircular ||
+        this.options.dereference.dereferenceOpts?.continueOnError) &&
       !ancestorsLineage.includesCycle(referencedElement)
     ) {
       // append referencing reference to ancestors lineage
@@ -452,14 +534,19 @@ class OpenAPI2DereferenceVisitor {
         reference,
         namespace: this.namespace,
         indirections: [...this.indirections],
-        options: this.options,
+        options: this.getNestedVisitorOptions(referencingElement),
         refractCache: this.refractCache,
         ancestors: ancestorsLineage,
       });
-      referencedElement = await visitAsync(referencedElement, visitor, {
-        keyMap,
-        nodeTypeGetter: getNodeType,
-      });
+      try {
+        referencedElement = await visitAsync(referencedElement, visitor, {
+          keyMap,
+          nodeTypeGetter: getNodeType,
+        });
+      } catch (error) {
+        this.indirections.pop();
+        return this.handleDereferenceError(error, referencingElement);
+      }
 
       // remove referencing reference from ancestors lineage
       directAncestors.delete(referencingElement);
@@ -539,15 +626,28 @@ class OpenAPI2DereferenceVisitor {
       return false;
     }
 
-    const reference = await this.toReference(toValue(referencingElement.$ref));
+    let reference: Reference;
+
+    try {
+      reference = await this.toReference(toValue(referencingElement.$ref));
+    } catch (error) {
+      return this.handleDereferenceError(error, referencingElement);
+    }
+
     const $refBaseURI = url.resolve(retrievalURI, toValue(referencingElement.$ref));
 
     this.indirections.push(referencingElement);
 
     const jsonPointer = URIFragmentIdentifier.fromURIReference($refBaseURI);
+    let referencedElement: Element;
 
-    // possibly non-semantic fragment
-    let referencedElement = evaluate<Element>(reference.value.result, jsonPointer);
+    try {
+      // possibly non-semantic fragment
+      referencedElement = evaluate<Element>(reference.value.result, jsonPointer);
+    } catch (error) {
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
+    }
     referencedElement.id = identityManager.identify(referencedElement);
 
     /**
@@ -574,14 +674,18 @@ class OpenAPI2DereferenceVisitor {
 
     // detect direct or circular reference
     if (referencingElement === referencedElement) {
-      throw new ApiDOMError('Recursive Reference Object detected');
+      const error = new ApiDOMError('Recursive Reference Object detected');
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
     }
 
     // detect maximum depth of dereferencing
     if (this.indirections.length > this.options.dereference.maxDepth) {
-      throw new MaximumDereferenceDepthError(
+      const error = new MaximumDereferenceDepthError(
         `Maximum dereference depth of "${this.options.dereference.maxDepth}" has been exceeded in file "${this.reference.uri}"`,
       );
+      this.indirections.pop();
+      return this.handleDereferenceError(error, referencingElement);
     }
 
     // detect second deep dive into the same fragment and avoid it
@@ -589,8 +693,12 @@ class OpenAPI2DereferenceVisitor {
       reference.refSet!.circular = true;
 
       if (this.options.dereference.circular === 'error') {
-        throw new ApiDOMError('Circular reference detected');
-      } else if (this.options.dereference.circular === 'replace') {
+        const error = new ApiDOMError('Circular reference detected');
+        this.indirections.pop();
+        return this.handleDereferenceError(error, referencingElement);
+      }
+
+      if (this.options.dereference.circular === 'replace') {
         const refElement = new RefElement(referencedElement.id, {
           type: 'json-reference',
           uri: reference.uri,
@@ -622,7 +730,8 @@ class OpenAPI2DereferenceVisitor {
       (isExternalReference ||
         isNonEntryDocument ||
         isJSONReferenceElement(referencedElement) ||
-        shouldDetectCircular) &&
+        shouldDetectCircular ||
+        this.options.dereference.dereferenceOpts?.continueOnError) &&
       !ancestorsLineage.includesCycle(referencedElement)
     ) {
       // append referencing reference to ancestors lineage
@@ -632,14 +741,19 @@ class OpenAPI2DereferenceVisitor {
         reference,
         namespace: this.namespace,
         indirections: [...this.indirections],
-        options: this.options,
+        options: this.getNestedVisitorOptions(referencingElement),
         refractCache: this.refractCache,
         ancestors: ancestorsLineage,
       });
-      referencedElement = await visitAsync(referencedElement, visitor, {
-        keyMap,
-        nodeTypeGetter: getNodeType,
-      });
+      try {
+        referencedElement = await visitAsync(referencedElement, visitor, {
+          keyMap,
+          nodeTypeGetter: getNodeType,
+        });
+      } catch (error) {
+        this.indirections.pop();
+        return this.handleDereferenceError(error, referencingElement);
+      }
 
       // remove referencing reference from ancestors lineage
       directAncestors.delete(referencingElement);