+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/swagger-api/apidom/compare/v1.0.2...v1.1.0) (2025-12-19)
+
+### Bug Fixes
+
+- **ns-asyncapi-3:** dereference schema external docs ([#5078](https://github.com/swagger-api/apidom/issues/5078)) ([a1fa740](https://github.com/swagger-api/apidom/commit/a1fa740ca281a6193bf1d1255924e50e31e10666))
+
+### Features
+
+- add option to continue on error when validating references ([#5070](https://github.com/swagger-api/apidom/issues/5070)) ([eedfe75](https://github.com/swagger-api/apidom/commit/eedfe755a601a4af7383d8d31fdb559b913a5e2e))
+
 ## [1.0.2](https://github.com/swagger-api/apidom/compare/v1.0.1...v1.0.2) (2025-12-16)
 
 ### Reverts